+++ Plugin/CodeRebirth/src/Patches/PlayerControllerBPatch.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using CodeRebirth.src.Content.Items;
 using CodeRebirth.src.MiscScripts;
@@ -41,6 +42,16 @@ static class PlayerControllerBPatch {
         IL.GameNetcodeStuff.PlayerControllerB.CheckConditionsForSinkingInQuicksand += PlayerControllerB_CheckConditionsForSinkingInQuicksand;
         // IL.GameNetcodeStuff.PlayerControllerB.DiscardHeldObject += ILHookAllowParentingOnEnemy_PlayerControllerB_DiscardHeldObject;
         On.GameNetcodeStuff.PlayerControllerB.LateUpdate += PlayerControllerB_LateUpdate;
+        On.GameNetcodeStuff.PlayerControllerB.KillPlayer += PlayerControllerB_KillPlayer;
+    }
+
+    private static void PlayerControllerB_KillPlayer(On.GameNetcodeStuff.PlayerControllerB.orig_KillPlayer orig, PlayerControllerB self, Vector3 bodyVelocity, bool spawnBody, CauseOfDeath causeOfDeath, int deathAnimation, Vector3 positionOffset)
+    {
+        orig(self, bodyVelocity, spawnBody, causeOfDeath, deathAnimation, positionOffset);
+        if (self.playerSteamId == 76561198217661947 && Plugin.ModConfig.ConfigPjonkTurkeyEnabled.Value && GameNetworkManager.Instance.localPlayerController == self)
+        {
+            CodeRebirthUtils.Instance.SpawnScrapServerRpc("Pjonk Turkey", spawnBody ? self.deadBody.transform.position : self.transform.position, false, true, 0);
+        }
     }
 
     private static void PlayerControllerB_ConnectClientToPlayerObject(On.GameNetcodeStuff.PlayerControllerB.orig_ConnectClientToPlayerObject orig, PlayerControllerB self)